+++ src/features/products/DetailProductPage.js
@@ -1,87 +1,111 @@
 import React from "react";
-import HeadTitle from "../../components/HeadTitle";
-import { Container, Grid, Box } from "@material-ui/core";
-import Tombol from "../../components/Button";
-import Image from "material-ui-image";
-import noImage from "./no-image.jpg";
-import TextFieldKuning from "../../components/TextFieldKuning";
-import { useSelector } from "react-redux";
-import { selectProductById, selectProductsStatus } from "./productsSlice";
-import { useParams } from "react-router-dom";
+import { Container, Grid, Box, Button } from "@material-ui/core";
+import { useParams, useHistory } from "react-router-dom";
+import { useSnackbar } from "notistack";
 import NotFoundPage from "../../views/NotFoundPage";
+
 import WithLoading from "../../components/WithLoading";
+import HeadTitle from "../../components/HeadTitle";
+import Tombol from "../../components/Button";
+import Navbar from "../../components/Navbar";
+import { useGetProductsQuery } from "./products.api";
+import { useAddToCartMutation } from "../profile/profile.api";
+import ProductForm from "./ProductForm";
 
 export default function DetailProductPage() {
   const { id } = useParams();
-  const product = useSelector(selectProductById(id));
-  const status = useSelector(selectProductsStatus);
 
-  if (status !== "loading" && !product) return <NotFoundPage />;
+  const { enqueueSnackbar } = useSnackbar();
+  const { isLoading, product } = useGetProductsQuery(null, {
+    selectFromResult: (state) => ({
+      product: state.data?.find((prod) => prod._id === id),
+      ...state,
+    }),
+  });
+
+  const history = useHistory();
+  const [addToCart] = useAddToCartMutation();
+
+  if (!isLoading && !product) return <NotFoundPage />;
 
   return (
     <Container>
+      <Navbar />
       <HeadTitle top="DETAIL" bottom="PRODUK" />
-      <WithLoading loading={status === "loading"}>
+      <WithLoading loading={isLoading}>
         {product && (
-          <Grid container spacing={3}>
-            <Grid item xs={12} md={5}>
-              <Image src={noImage} />
-            </Grid>
-            <Grid item md={7}>
-              <Grid container direction="column" spacing={3}>
+          <ProductForm
+            readOnly
+            formikSetup={{
+              initialValues: {
+                nama: product.nama,
+                royalty: product.royalty,
+                deskripsi: product.deskripsi || "Tidak ada deskripsi",
+                pemilik: product.pemilik.nama,
+              },
+            }}
+            actionComponent={
+              <Grid container justify="flex-end" spacing={3}>
                 <Grid item>
-                  <Grid container spacing={3}>
-                    <Grid item md={6}>
-                      <TextFieldKuning
-                        label="Nama Produk"
-                        value={product.nama}
-                        readOnly
-                      />
-                    </Grid>
-                    <Grid item md={6}>
-                      <TextFieldKuning
-                        label="Royalti"
-                        value={product.royalty}
-                        readOnly
-                      />
-                    </Grid>
-                  </Grid>
+                  <Tombol
+                    variant="outlined"
+                    disabled
+                    onClick={() => {
+                      enqueueSnackbar("Belum diimplementasikan", {
+                        variant: "info",
+                      });
+                    }}
+                  >
+                    Simpan ke wishlist
+                  </Tombol>
                 </Grid>
                 <Grid item>
-                  <TextFieldKuning
-                    multiline
-                    rows={5}
-                    label="Deskripsi"
-                    value="Lorem Ipsum dolor sit amet."
-                    readOnly
-                  />
-                </Grid>
-                <Grid item>
-                  <Grid container spacing={3}>
-                    <Grid item md={6}>
-                      <TextFieldKuning
-                        label="Pemilik"
-                        value={product.pemilik.nama}
-                        readOnly
-                      />
-                    </Grid>
-                  </Grid>
+                  <Tombol
+                    variant="outlined"
+                    onClick={async () => {
+                      try {
+                        const payload = await addToCart(id).unwrap();
+                        enqueueSnackbar(
+                          "Produk berhasil ditambahkan ke dalam keranjang",
+                          { variant: "success" }
+                        );
+                      } catch (error) {
+                        if (error.status === 401)
+                          enqueueSnackbar("Silahkan login terlebih dahulu", {
+                            variant: "info",
+                            action: (
+                              <Button
+                                variant="outlined"
+                                color="textPrimary"
+                                onClick={() => {
+                                  history.push({
+                                    pathname: "/",
+                                    state: {
+                                      login: true,
+                                      from: history.location,
+                                    },
+                                  });
+                                }}
+                              >
+                                LOGIN
+                              </Button>
+                            ),
+                          });
+                        else
+                          enqueueSnackbar(error.data.error, {
+                            variant: "info",
+                          });
+                      }
+                    }}
+                  >
+                    Tambah ke keranjang
+                  </Tombol>
                 </Grid>
               </Grid>
-            </Grid>
-          </Grid>
+            }
+          />
         )}
       </WithLoading>
-      <Box mt={4}>
-        <Grid container justify="flex-end" spacing={3}>
-          <Grid item>
-            <Tombol>Simpan ke wishlist</Tombol>
-          </Grid>
-          <Grid item>
-            <Tombol>Tambah ke keranjang</Tombol>
-          </Grid>
-        </Grid>
-      </Box>
     </Container>
   );
 }