+++ src/features/profile/LoginDialog/MainForm.js
@@ -15,6 +15,7 @@ function MainForm({ registerMode, ...props }) {
       {registerMode && (
         <Grid item xs={12}>
           <Field
+            autoFocus
             component={TextField}
             name="nama"
             className={classes.textFieldCustom}