+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.0] - 2021-06-22
+
+> [backend]: 0.4.0
+
+### Added
+
+- Add, edit, and delete product
+- Add and remove product from carts
+
 ## [0.4.0] - 2021-06-21
 
 > [backend]: 0.3.0